finished adding authentication anonymously.
It seems like the anonymous authentication is working, but now
I need to set up a landing widget to control the state of the app.
Firebase seems to have received an user that logged in.

There is an error about a gradle exception. It doesn't seem to interrupt
the app from working and I couldn't find a solution.