implementing stream to manage states instead of using callbacks
This is the implementation of the a Stream that will handle user changes
any time. This makes it easier to handle state without having to link
every single state in the app with callbacks to widgets. This does
NOT remove the callback implementation. Right now both, callbacks
and stream, are working. There is a variable that is unused but that
will be removed when the callbacks are removed and replaced with the
stream.