Landing is not stateless, implements streams and StreamBuilder
By getting rid of callbacks, is no longer necessary to keep track of
the state in a stateful widget as we will be using a stream for
authentication. This means that the Landing page no longer needs to be
stateful and can transition into stateless by using a StreamBuilder
widget to listen to the stream. This simplifies the code as we can
delete all methods designed to keep state and to do callbacks.