project name change and google sign in
There is a name change. When I created this app, I named it
"just serve" but I did not know that there is a service app called
exactly that. So I'm changing every place I can to the new placeholder
which is "lose yourself." There are parts of the project that can no
longer be changed and this I had to choose between creating a new project
and copying what I have, or to keep this project and rename the public
text so that version control is maintained. I am going for the latter
approach.

Google sign-in has been enabled. The next step is enabling Google sign
out as the current sign out method only signs out of firebase but not
from Google.