Modifying the form to make sure it is working properly and good UX
There were quite a few changes. Now there are buttons that direct the
user to the next box in the keyboard. This is achieved by using
FocusNodes. There are messages that appear when the user has
tried to submit the form and the text is invalid

Added a class to handle validation for the strings which will make it
simpler to maintain in the future. Right now there is an issue with
overflow of pixels when the error text appears. This needs to be fixed.