avoid log in with keyboard when not in valid state. Fix overflow
The app had some logic to disable the log in button if the email and
password fields were not valid. However, the user was still able to
log in using enter and the log in method would kick in. This has been
fixed now.

The app was fixed which led to some overflow problems. The widget has
been set to be scrollable in order to avoid the overflow problem.