Platform aware widgets
This ends the platform aware widgets. As part of this
section, there were a lot of changes in dependencies
and what not. This is because I started developing
this app on a different computer and I was having
problems trying to log in with Google. So I changed
the dependencies hoping that something would fix
the behavior. I realized that I was generating the
wrong type of SHA1 key for firebase, but at that point I
had already changed too many things and my app is now
working. So I'm deciding to leave all the changes as
things are currently working.