Pass AuthProvider instead of method chaining everything
This is a big commit, an Auth Provider class is created
and it replaces all the method chaining that had been
done in order to pass the auth from the main page. The
Auth object is created once in main and then accessed
with the context all over the app.