make sign in page stateful, use dispose, set colors
The sign in page now has a widget that shows whether
there is a request sent to firebase that comes from
using Google authentication. The color of the buttons
has been set to the same color whether these are
disabled or not.

Calling dispose on focus nodes and controllers on
the email sign in page to make sure these widgets
are disposed of correctly.