using a bloc and a model in the email sign in form
This is a big change. All the business logic has
been moved to the bloc class and model classes.
The bloc class modifies the model, the model
gives access to the variables. The bloc handles the
authentication with firebase. The only logic
remaining in the email sign in form is UI related.
This way concerns are separated and it's easier to
test the business logic.