Firestore DB is now working.
There is a button on the projects page that has a +
button to write to the database. The database that
will be used is firestore db and it allows for real time
reading through a stream. The paths are not all the way
ready but data has been able to be read from the db
and it is showing in the ListView in the projects.

A lot has changed. Instead of having a home page,
there is a projects page. There is a class to hold all
the paths to write to the database and the sign in page
has been modifies accordingly. There is a class to use
firestore as a service in order to separate the concerns.

The code might have a problem which prevents certain
changes to be 'hot reloaded' and I will have to
look more into it.

The rules to the database need to be changed if any of
the paths change.