Add projects page using text form fields with a key
The form for adding a project has been added.
The form does not send data to firestore right now,
but that is the next step. Right now the form saves
and validates data. No fields are validated as of right
now as I haven't decided how much validation I want to
add to this form. This page is linked with the projects
page and it is accessed through the '+' button in the
projects page.