Finished create project form and unique project id
The form to add projects is now working and using
focus nodes to make navigation simpler for the user.

The projects now have a unique id that's created from
the time that the project is created down to the
microsecond. In the unlikely event that two projects
are created simultaneously at the exact time, a
random number is appended at the end of this to
avoid a problem with two different projects that
will try to write to each other.