change the UI for showing projects and enable editing a project
The UI is better at showing projects. They can be tapped and edited
now. There is a need to get back to some fixes that are not
necessary but are nice. I left TODO comments.

The database and firestore service classes have been modified to
enable editing without many changes.