CRUD operations finished
The app is able to add, edit, read, and delete projects
from the database. There is still a need to show a list
of all the public projects and add validation for the
fields. I think the same page can be reused to show the
list of projects, as the only different things are
that projects should not be deleted from there, and
that they cannot be edited from there.