Navigation bar and scaffold for other pages
Instead of keeping the state of the project page, now there are
two different widgets to represent personal and public projects.
These seemed different enough that recycling the code would just
make it too confusing to read so I will err on the side of
readability. There are things that still need to be done.

The details page should not use forms and probably just be
a stateless widget. Also, the submit method for personal
projects could use some cleaning.

The floating button in the personal project page has been added
to the task bar as the floating button and the cupertino
tab don't play along very well. The log out button has been moved
to the account page.